Support different brands of Colored Lights.
Change AqualinkLightType from a simple Enum to a full-featured dataclass.
Replace the AqualinkLightEffect Enum with a string to string mapping (so that we can use spaces and punctuation in the display names.)
Add the light types and effect names as retrieved from a REV T.2 AquaLink device.
Allow retrieving the supported light effects for the current device.
Allow setting the light effect by name or by number.
Fix a few type warnings.